+++ src/lib/components/Hero.svelte
@@ -22,23 +22,24 @@
     <link rel="preload" href="/{info.avatar}" as="image">
 </svelte:head>
 
-<section class="card-shimmer z-0 grid grid-cols-1 lg:grid-cols-2 bg-surface-100-800-token border-t border-b md:border border-surface-200-700-token shadow-sm overflow-hidden md:rounded-container rounded-none">
-    <div class="relative p-6 sm:p-8 row-span-2">
-        <div class="relative aspect-square" style="padding-bottom: 33.33%;">
-            {#if !imageLoaded}
-                <div class="placeholder absolute inset-0 animate-pulse rounded-container h-full bg-surface-50-900-token"></div>
-            {:else}
-                <!-- Apply fade transition to the image -->
-                <div class="absolute inset-0 rounded-container" style="background: url('{info.avatar}') no-repeat center / cover;" in:fade={{duration: 100}}></div>
-            {/if}
-        </div>
-    </div>
-    <div class="flex flex-col p-6 md:p-8 md:pt-0 lg:pl-0 pt-0 lg:pt-8 row-span-2 overflow-auto">
-        <p class="text-lg font-normal flex-grow">   
+<section class="card-shimmer z-0 grid grid-cols-1 md:grid-cols-3 bg-surface-100-800-token border-t border-b md:border border-surface-200-700-token shadow-sm overflow-hidden md:rounded-container rounded-none md:gap-4">
+    <div class="flex flex-col p-6 md:p-8 md:pr-0 pt-0 overflow-auto col-span-2 row-start-2 md:row-start-1">
+        <h2 class="text-2xl font-semibold pb-4">{info.name}</h2>
+        <p class="text-base font-normal flex-grow">   
             {@html info.bio.html}
         </p>
         <div class="pt-8 text-base">
             <Socials />
+        </div>
+    </div>
+    <div class="relative p-6 lg:p-8 block md:flex-col md:justify-center"> <!-- Added flex utilities here -->
+        <div class="relative aspect-square mx-auto h-36 md:h-auto">
+            {#if !imageLoaded}
+                <div class="placeholder absolute inset-0 animate-pulse rounded-[3rem] md:rounded-[4.5rem] h-full bg-surface-100-800-token"></div>
+            {:else}
+                <!-- Apply fade transition to the image -->
+                <div class="absolute inset-0 rounded-[3rem] md:rounded-[4.5rem]" style="background: url('{info.avatar}') no-repeat center / cover;" in:fade={{duration: 100}}></div>
+            {/if}
         </div>
     </div>
 </section>