+++ src/routes/+layout.svelte
@@ -27,6 +27,7 @@
 		document.getElementById('page')?.scrollTo(0, 0);
 	});
 
+	let menuItems = info.menuItems;
 	let link;
 	let isDark = false;
 	let logo;
@@ -91,10 +92,9 @@
 		<div class="w-fit">
 			<Menu>
 				<svelte:fragment slot="small-screens">
-					<MenuButton href="/books">Books</MenuButton>
-					<MenuButton href="/projects">Projects</MenuButton>
-					<MenuButton href="/changelog">Changelog</MenuButton>
-					<MenuButton href="https://github.com/koenraijer/koenraijer.2.0">Source</MenuButton>
+					{#each menuItems as item}
+						<MenuButton href={item.href}>{item.title}</MenuButton>
+					{/each}
 					<hr>
 					<div class="flex row-nowrap justify-between gap-x-2">
 						<ThemeToggle />
@@ -102,10 +102,9 @@
 					</div>
 				</svelte:fragment>
 				<svelte:fragment slot="large-screens">
-					<A href="/books" classes="whitespace-nowrap hover:underline font-normal text-lg text-surface-900-50-token">Books</A>
-					<A href="/projects" classes="whitespace-nowrap hover:underline font-normal text-lg text-surface-900-50-token">Projects</A>
-					<A href="/changelog" classes="whitespace-nowrap hover:underline font-normal text-lg text-surface-900-50-token">Changelog</A>
-					<A href="https://github.com/koenraijer/koenraijer.2.0" classes="whitespace-nowrap hover:underline font-normal text-lg text-surface-900-50-token">Source</A>
+					{#each menuItems as item}
+						<A href={item.href} classes="whitespace-nowrap hover:underline font-normal text-base text-surface-900-50-token">{item.title}</A>
+					{/each}
 					<ThemeToggle />
 				</svelte:fragment>
 			</Menu>