+++ src/routes/books/[slug]/+page.server.js
@@ -7,7 +7,6 @@ export async function load({ params, fetch }) {
         // Match slug to slug in data
         const book = data.find(book => book.slug === params.slug)
         
-        console.log("+page.server.js data", book)
         return {
             book
         }