+++ src/routes/changelog/CHANGELOG.md
@@ -1,6 +1,7 @@
+### May
+- Added [Projects](/projects), improved UI of hero and [Books](/books)
 ## 2024
 ### January
-- sth
 - Added post [Links and resources](https://www.koenraijer.io/links-and-resources)
 - Changed link component so icons for external links are added automatically, changed menus
 - Added lazy loading to [Books](/books) based on intersection observer API ([src](https://alex-schnabl.medium.com/lazy-loading-images-and-components-in-svelte-and-sveltekit-using-typescript-6a8443bb9479))